chore: rename GalleryItemCardList component
