+++ components/shared/gallery/GalleryItemCardList.vue
